[nn] split embedding tests from test_nn
